% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_mape_municipios}
\alias{download_mape_municipios}
\title{Function to download the main data file}
\usage{
download_mape_municipios(quiet = FALSE)
}
\arguments{
\item{quiet}{Logical. If \code{TRUE}, suppress messages and progress bar. Default is \code{FALSE}.}
}
\value{
A tibble with the main data file
}
\description{
The \code{download_mape_municipios} function downloads the main data file from the
\href{https://osf.io/3yka9/}{OSF} main repository. It takes no arguments, downloads the
file (named as \code{mape_municipios.Rda} in the working directory), and returns a tibble
with the main data file.
}
\examples{
\dontrun{
mape_muns <- download_mape_municipios()
}
}
