% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPEmunicipios-package.R
\docType{package}
\name{MAPEmunicipios-package}
\alias{MAPEmunicipios}
\alias{MAPEmunicipios-package}
\title{MAPEmunicipios: Functions to Download and Use the MAPEmunicipios Dataset}
\description{
Provides a set of functions to download and use the MAPEmunicipios dataset. The dataset contains detailed information of Brazilian municipalities, covering more than 451 variables across 31 original sources.
}
\author{
\strong{Maintainer}: Fernando Meireles \email{fernando.meireles@iesp.uerj.br} (\href{https://orcid.org/0000-0002-7027-2058}{ORCID})

Authors:
\itemize{
  \item Bruno Schaefer \email{bruno.schaefer@iesp.uerj.br} (\href{https://orcid.org/0000-0002-5127-7240}{ORCID})
  \item Carlos Freitas \email{carlosfreitas@iesp.uerj.br} (\href{https://orcid.org/0000-0002-2493-8154}{ORCID})
  \item Tomás Paixão Borges \email{tomasborges@iesp.uerj.br} (\href{https://orcid.org/0000-0002-5276-6636}{ORCID})
  \item Olga Caldas \email{olgasilva@iesp.uerj.br}
  \item Luisa Barreto \email{luisafernandes@iesp.uerj.br} (\href{https://orcid.org/0009-0001-2488-5806}{ORCID})
  \item Frederico Gomes Filho \email{fredericofilho@iesp.uerj.br} (\href{https://orcid.org/0009-0004-2217-8077}{ORCID})
  \item Larissa Mendes \email{larissamendes@iesp.uerj.br} (\href{https://orcid.org/0009-0004-2217-8077}{ORCID})
}

}
\keyword{internal}
